Create AdditionalPropertiesAllowed subdataclass
