+++ core/dbt/contracts/graph/parsed.py
@@ -12,7 +12,8 @@ import dbt.clients.jinja
 import dbt.flags
 from dbt.contracts.graph.unparsed import (
     UnparsedNode, UnparsedMacro, UnparsedDocumentationFile, Quoting,
-    UnparsedBaseNode, FreshnessThreshold#, ExternalTable
+    UnparsedBaseNode, FreshnessThreshold, ExternalTable, 
+    AdditionalPropertiesAllowed
 )
 from dbt.contracts.util import Replaceable
 from dbt.logger import GLOBAL_LOGGER as logger  # noqa
@@ -56,7 +57,7 @@ register_pattern(Severity, insensitive_patterns('warn', 'error'))
 
 @dataclass
 class NodeConfig(
-    ExtensibleJsonSchemaMixin, Replaceable, MutableMapping[str, Any]
+    AdditionalPropertiesAllowed, Replaceable, MutableMapping[str, Any]
 ):
     enabled: bool = True
     materialized: str = 'view'
@@ -72,25 +73,6 @@ class NodeConfig(
     @property
     def extra(self):
         return self._extra
-
-    @classmethod
-    def from_dict(cls, data, validate=True):
-        self = super().from_dict(data=data, validate=validate)
-        keys = self.to_dict(validate=False, omit_none=False)
-        for key, value in data.items():
-            if key not in keys:
-                self._extra[key] = value
-        return self
-
-    def to_dict(self, omit_none=True, validate=False):
-        data = super().to_dict(omit_none=omit_none, validate=validate)
-        data.update(self._extra)
-        return data
-
-    def replace(self, **kwargs):
-        dct = self.to_dict(omit_none=False, validate=False)
-        dct.update(kwargs)
-        return self.from_dict(dct)
 
     @classmethod
     def field_mapping(cls):
@@ -470,7 +452,7 @@ class ParsedSourceDefinition(
     quoting: Quoting = field(default_factory=Quoting)
     loaded_at_field: Optional[str] = None
     freshness: Optional[FreshnessThreshold] = None
-    external: Optional[Dict[str, Any]] = None
+    external: Optional[ExternalTable] = None
     docrefs: List[Docref] = field(default_factory=list)
     description: str = ''
     columns: Dict[str, ColumnInfo] = field(default_factory=dict)
